backend/requirements.txt
alembic==1.7.5
amqp==5.0.9
anyio==3.5.0
asgiref==3.5.0
astroid==2.12.9
attrs==21.4.0
Authlib==1.1.0
bcrypt==3.2.0
billiard==3.6.4.0
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==38.0.1
cssselect==1.1.0
cssutils==2.3.0
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
emails==0.6
factory_boy==3.3.1
Faker==27.4.0
fastapi==0.75.0
fastapi-pagination==0.9.1
flower==1.0.0
google-auth==2.13.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httptools==0.3.0
httpx==0.23.0
humanize==4.0.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.0.3
kombu==5.2.3
lazy-object-proxy==1.7.1
lxml==4.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
platformdirs==2.5.2
pluggy==1.0.0
premailer==3.10.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pylint==2.15.2
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
tomli==2.0.1
tomlkit==0.11.4
tornado==6.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0.post1
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
wrapt==1.13.3
zipp==3.7.0
